requirements.txt
asgiref==3.7.2
attrs==23.1.0
black==23.10.1
click==8.1.7
Django==4.2.6
django-js-asset==2.1.0
django-mptt==0.15.0
djangorestframework==3.14.0
drf-spectacular==0.26.5
exceptiongroup==1.1.3
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
pluggy==1.3.0
pytest==7.4.3
pytest-django==4.6.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.6
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
uritemplate==4.1.1
